Tweak camera resolution and FPS
According to some random comment I saw on Discord this is an optimal
setup. This could of course be changed though I don't see a nice way to
make it Tunable.

Thanks to the magic of quadratics this reduces the bandwidth by
a factor of 8 (!) ignoring compression. It'll look crap, sure, but you
don't exactly need 1080p to make out giant bright yellow cubes.