Use Jaci's EncoderFollower for profile control
This is a reversion to open-loop control (as we think of it, technically
it's still closed-loop because there's no human input but shut up) in
the name of "I think it might work". Preliminary tests showed that this
caused the robot to have a mild seizure before slamming into a wall, but
progress should be forthcoming with more testing.

This has  the benefit of us not needing to figure out our own following
algorithm so that should help.